cleanup(core): move wrapArgIntoQuotesIfNeeded into shell-quoting utils
## Current Behavior
Helper lives privately inside run-commands.impl.ts with no direct unit tests.

## Expected Behavior
Helper exported from utils/shell-quoting.ts (alongside needsShellQuoting / isAlreadyQuoted) so other executors can reuse it, with unit tests covering JSON values, --key=value form, plain args, and already-quoted args. No behavior change. Also folds in an incidental pretty-quick reformat of a nested ternary elsewhere in run-commands.impl.ts.
