+++ packages/nx/src/utils/shell-quoting.ts
@@ -39,3 +39,30 @@ export function isAlreadyQuoted(str: string): boolean {
       (str[0] === '"' && str[str.length - 1] === '"'))
   );
 }
+
+/**
+ * Wrap a CLI arg in double quotes when it contains shell metacharacters,
+ * so it survives being passed to `spawn(..., { shell: true })`. For
+ * `--key=value` form, only the value side is quoted.
+ */
+export function wrapArgIntoQuotesIfNeeded(arg: string): string {
+  if (arg.includes('=')) {
+    const eqIndex = arg.indexOf('=');
+    const key = arg.substring(0, eqIndex);
+    const value = arg.substring(eqIndex + 1);
+    if (
+      key.startsWith('--') &&
+      needsShellQuoting(value) &&
+      !isAlreadyQuoted(value)
+    ) {
+      const escaped = value.replace(/"/g, '\\"');
+      return `${key}="${escaped}"`;
+    }
+    return arg;
+  } else if (needsShellQuoting(arg) && !isAlreadyQuoted(arg)) {
+    const escaped = arg.replace(/"/g, '\\"');
+    return `"${escaped}"`;
+  } else {
+    return arg;
+  }
+}
